Move launchpad.py to a Python module (launchpad_py) and add a setup.py installer
