+++ setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup
+
+setup(
+    name = "launchpad_py",
+    version = "0.9.0",
+    description = "A Novation Launchpad control suite for Python",
+    author = "ASkr (FMMT666)",
+    license = "CC BY 4.0",
+    keywords = "novation launchpad midi",
+    url = "https://github.com/FMMT666/launchpad.py",
+    packages = ["launchpad_py"],
+)
